EDA and Prediction with best score results
This file consists of detailed modelling of the dataset with Hot Rod Encoding tht will give us the bet results. Each parameters have been explained properly and how it has been effecting the process of prediction. 
With proper inputs and entering the values, we can predict if the customer will keep using the services of the company or will company will lose the customer.